+++ src/components/Header.tsx
@@ -3,7 +3,6 @@ import {
   IconFileStack,
   IconGlobe,
   IconHome,
-  IconLanguage,
   IconMenu,
   IconNetwork,
   IconPalette,
@@ -11,8 +10,8 @@ import {
   IconSettings,
 } from '@tabler/icons-solidjs'
 import { For, ParentComponent, Show, createSignal } from 'solid-js'
-import { Button, LogoText } from '~/components'
-import { LANG, ROUTES, themes } from '~/constants'
+import { LogoText } from '~/components'
+import { ROUTES, themes } from '~/constants'
 import { useI18n } from '~/i18n'
 import { setCurTheme } from '~/signals'
 
@@ -62,7 +61,7 @@ const ThemeSwitcher = () => (
 )
 
 export const Header = () => {
-  const { t, locale } = useI18n()
+  const [t] = useI18n()
   const navs = () => [
     {
       href: ROUTES.Overview,
@@ -152,16 +151,6 @@ export const Header = () => {
 
       <div class="navbar-end">
         <div class="flex items-center gap-2">
-          <Button
-            class="btn-circle btn-secondary btn-sm"
-            onClick={() => {
-              const curLocale = locale()
-
-              locale(curLocale === LANG.EN ? LANG.ZH : LANG.EN)
-            }}
-            icon={<IconLanguage />}
-          />
-
           <ThemeSwitcher />
         </div>
       </div>