+++ src/components/ProxiesSettingsModal.tsx
@@ -21,7 +21,7 @@ import {
 export const ProxiesSettingsModal: Component<{
   ref?: (el: HTMLDialogElement) => void
 }> = (props) => {
-  const { t } = useI18n()
+  const [t] = useI18n()
 
   return (
     <Modal