+++ src/i18n/en.ts
@@ -105,4 +105,7 @@ export default {
   tunModeStack: 'TUN Mode Stack',
   tunDeviceName: 'TUN Device Name',
   interfaceName: 'Interface Name',
+  en: 'English',
+  zh: 'Chinese',
+  port: '{{ name }} Port',
 }