+++ src/i18n/zh.ts
@@ -107,4 +107,7 @@ export default {
   tunModeStack: 'TUN 模式堆栈',
   tunDeviceName: 'TUN 设备名称',
   interfaceName: '接口名称',
+  en: '英文',
+  zh: '中文',
+  port: '{{ name }} 端口',
 } satisfies Dict