+++ src/pages/Logs.tsx
@@ -15,7 +15,7 @@ import {
   getFilteredRowModel,
   getSortedRowModel,
 } from '@tanstack/solid-table'
-import { For, Index, createEffect, createMemo, createSignal } from 'solid-js'
+import { For, Index, createEffect, createSignal } from 'solid-js'
 import { twMerge } from 'tailwind-merge'
 import { Button, LogsSettingsModal } from '~/components'
 import { LOG_LEVEL } from '~/constants'
@@ -42,7 +42,8 @@ const fuzzyFilter: FilterFn<LogWithSeq> = (row, columnId, value, addMeta) => {
 export default () => {
   let logsSettingsModalRef: HTMLDialogElement | undefined
 
-  const { t } = useI18n()
+  const [t] = useI18n()
+
   let seq = 1
   const [logs, setLogs] = createSignal<LogWithSeq[]>([])
 
@@ -67,7 +68,7 @@ export default () => {
     storage: localStorage,
   })
 
-  const columns = createMemo<ColumnDef<LogWithSeq>[]>(() => [
+  const columns: ColumnDef<LogWithSeq>[] = [
     {
       header: t('sequence'),
       accessorFn: (row) => row.seq,
@@ -102,7 +103,7 @@ export default () => {
       header: t('payload'),
       accessorFn: (row) => row.payload,
     },
-  ])
+  ]
 
   const table = createSolidTable({
     filterFns: {
@@ -120,7 +121,7 @@ export default () => {
       return logs()
     },
     sortDescFirst: true,
-    columns: columns(),
+    columns,
     onGlobalFilterChange: setGlobalFilter,
     onSortingChange: setSorting,
     globalFilterFn: fuzzyFilter,